Paginación

+++ src/pages/Private.js
@@ -6,18 +6,24 @@ import Card from '../components/Card'
 import Paginador from '../components/Paginador'
 
 class Private extends Component {
-  limite = 5
-  state= {
-    houses: [],
-    paginador: {
-      actual: 1,
-      offset:0
-    },
-    numHouses: ''
-  }
 
-  componentDidMount(){
-    houseBackendService.getAllHouses()
+state= {
+  houses: [],
+  paginador: {
+    page: 1,
+    per_page: 5
+  },
+  numHouses: '',
+  message: 'Vivienda Eliminada',
+  success: false
+}
+
+componentDidMount(){
+  this.loadHouses()
+}
+  loadHouses(){
+    const {paginador} = this.state
+    houseBackendService.getAllHouses(paginador.page, paginador.per_page)
     .then(response => {
       // console.log(response)
       this.setState({
@@ -26,28 +32,40 @@ class Private extends Component {
       })
     })
   }
+  onSuccessfulSubmit = ()=> {
+    this.setState({
+      success: true
+    }, () => {
+      setTimeout(()=>{
+        this.setState({
+          success: false
+        })
+      }, 1500)
+    })
+  }
 
   pagePrevious = () => {
     this.setState({
       paginador:{
-        offset: this.state.paginador.offset - this.limite,
-        actual: this.state.paginador.actual - 1
+        per_page: this.state.paginador.per_page,
+        page: this.state.paginador.page - 1,
       }
-    })
+    }, this.loadHouses)
   }
 
   pageNext = () => {
     this.setState({
       paginador:{
-        offset: this.state.paginador.offset + this.limite,
-        actual: this.state.paginador.actual + 1
+        per_page: this.state.paginador.per_page,
+        page: this.state.paginador.page + 1
       }
-    })
+    }, this.loadHouses)
   }
   handleDeleteClick = (id) => {
     const {houses} = this.state;
     houseBackendService.deleteOneHouse(id)
     .then(()=>{
+      this.onSuccessfulSubmit()
       const filterHouses = houses.filter((house)=> {
         return house._id !== id
       })
@@ -58,7 +76,7 @@ class Private extends Component {
   }
   
   render() {
-    const {houses, numHouses, paginador} = this.state
+    const {houses, numHouses, paginador, success, message} = this.state
     
     return (
         <div className="container d-flex">
@@ -76,8 +94,11 @@ class Private extends Component {
                     <Link to={`/houses/details/${house._id}`}><button type="button" className="btn w-80 btn-warning bar mb-4">Ver Datos Vivienda</button></Link>
                     <Link to={`/houses/update/${house._id}`}><button type="button" className="btn w-80 btn-warning bar mb-4">Modificar Datos Vivienda</button></Link>
                     <button type="button" className="btn w-80 btn-danger bar mb-4" onClick={()=> {
-                      this.handleDeleteClick(house._id)
-                    }}>Borrar</button>
+                      if(window.confirm('Confirme para eliminar vivienda')){
+                        this.handleDeleteClick(house._id)
+                          }
+                      }}>Borrar</button>
+                    { success ? <h4 className="bg-danger p-4 message">{message}</h4> : '' }
                   </div>
                   </div>
                   )
@@ -85,9 +106,9 @@ class Private extends Component {
 
             </div>
             <Paginador 
-              actual= {paginador.actual}
+              page= {paginador.page}
               numHouses = {numHouses}
-              limite={this.limite}
+              per_page={paginador.per_page}
               pagePrevious={this.pagePrevious}
               pageNext={this.pageNext}
             />
